spec 039 review: ANALYZE re-derives a table's projection, a check judged live, tidy across the tree
Two adversarial passes over spec 039, applied:
- schema_stale on a declared-list table named ANALYZE VIRTUAL TABLE, which re-derived only views
  and macros: CatalogRefreshSchema now re-derives a table's projection schema too - and carries
  the declared nullability marks and column comments through every re-derivation (it used to
  drop them for views and macros);
- a grant's items were judged against a view's STORED schema - exactly what may be stale - so a
  catalog-wide bare name that matched nothing any more went unreported; judged against the declared
  list, else what the source binds to now;
- the listing mark called a bare-word function (`who = current_user`) a missing column while the
  check was clean: SQL's bare-word functions and constants are not judged; and the mark is one
  grouped scan of information_schema.columns per listing instead of a correlated subquery per
  relation on every principal's listing;
- DROP MISSING COLUMNS AND MASKS is several catalog writes, so the order is what keeps a failure
  midway harmless: grants lose their mask items first (fewer items read less, never more), the
  object's list is cut after;
- the spec's vocabulary, refusal text and repair column now say what the code does (no
  mask_orphaned; a catalog grant's mask is never touched; object grants are re-granted through
  acl_grant_object, which has no ALTER form); duckdb_views() never carries the mark (a view has no
  declared list).

Tests: every problem of the vocabulary is produced at least once (rls_unchecked, schema_stale on a
table and on a view, mask_broken from an object grant, grant_column_missing for a catalog grant with
object `*`, expansion_stale in both halves with REFRESH PRUNE, reference_dangling for an end and for
a column, definition_broken for a macro template, a scalar alias and a view, source_missing on a
bare alias and a declared table), the binder error is pinned, the no-argument check walks a second
catalog with a broken object of its own, a computed column written as a bare-word function is not
marked. 233 assertions.

clang-tidy across our sources (`make tidy`, new: the repo's duckdb .clang-tidy over the compile
database, with the SDK sysroot macOS needs): by-value shared_ptr / function parameters made const
references, the deprecated DataChunk::SetValue/SetCardinality replaced, a use-after-move in the
rewriter's insert guard restructured (the guard now rides the first column outside the loop), an
int-typed year constant, `long long` casts dropped, a std::move into a const reference dropped, the
two deliberate process-wide caches and the thread-local seam marked NOLINT with their reasons.
Unused includes dropped where clang-include-cleaner and the compiler agree (`duckdb.hpp` and
`<arrow/table.h>` were transitively load-bearing and stay).

Whole test/sql (60 files), test-cpp, harness, test-flight, door e2e (postgres/ducklake/pair) and
integration green locally; `make tidy` clean.

+++ src/acl_policy.cpp
@@ -907,7 +907,7 @@ bool PolicyStore::SetSessionAuditLevel(const string &id, int8_t level) {
 int8_t PolicyStore::SessionAuditLevel(const string &handle) {
 	lock_guard<mutex> guard(lock);
 	auto entry = sessions.find(handle);
-	return entry == sessions.end() ? -1 : entry->second.audit_level;
+	return entry == sessions.end() ? static_cast<int8_t>(-1) : entry->second.audit_level;
 }
 
 vector<std::pair<string, int64_t>> PolicyStore::JwksAges() {
@@ -1184,6 +1184,7 @@ void PolicyStore::MapRole(const string &issuer, const string &source, const stri
 //! The exec-context seam (spec 050). Thread-local: the door sets it on the thread that calls
 //! Prepare, the rewriter reads it on that same thread during the parse inside that Prepare, and it
 //! is cleared before the exec lock is released - no other thread ever observes a value.
+// NOLINTNEXTLINE(cppcoreguidelines-avoid-non-const-global-variables)
 static thread_local ClientContext *temp_scan_context = nullptr;
 
 void SetTempScanContext(ClientContext *context) {
